requirements.txt
# Web / API
fastapi==0.115.6
uvicorn[standard]==0.32.1

# HTTP
requests==2.32.3

# PostgreSQL
psycopg2-binary==2.9.10

# Env vars
python-dotenv==1.0.1
